[neutil] Move doc generation to sub command.
This makes room for other subcommands.
